move tau embedding header files to interface folder
