+++ TauAnalysis/MCEmbeddingTools/interface/CollectionMerger.h
