+++ TauAnalysis/MCEmbeddingTools/interface/DoubleCollectionMerger.h
