+++ TauAnalysis/MCEmbeddingTools/interface/EmbeddingBeamSpotOnlineProducer.h
